finish accept process for transactions
