+++ Standard.md
@@ -2,6 +2,22 @@ Some transactions will not be accepted by miners unless they appear in a block.
 This is equivalent to the 'IsStandard' function in Bitcoin. This file dictates
 the rules for standard Sia transactions.
 
+------------------------
+-- Double Spend Rules --
+------------------------
+
+When two conflicting transactions are seen, the first transaction is the only
+one that is kept. If the blockchain reorganizes, the transaction that is kept
+is the transaction that was most recently in the blockchain. This is to
+discourage double spending, and enforce that the first transaction seen is the
+one that should be kept by the network.
+
+Transactions are currently included into blocks using a first-come first-serve
+algorithm. Eventually, transactions will be rejected if the fee does not meet a
+certain minimum. For the near future, there are no plans to prioritize
+transactions with substantially higher fees. Other mining software may take
+alternative approaches.
+
 -------------------------
 -- Storage Proof Rules --
 -------------------------
@@ -15,6 +31,11 @@ invalidated by a blockchain reorg - if the trigger block changes, the proof
 will be invalidated. Storage proofs can by any reorg, where standard
 transactions can only be invalidated by a doublespend (which requires a
 signature from the double spender).
+
+This also means that transaction pools will track multiple conflicting storage
+proofs. If there are two competing reorgs, it is in the best interest of the
+network to keep storage proofs for each reorg, because each proof will only be
+valid on one reorg.
 
 --------------------------
 -- Arbitrary Data Usage --
