+++ modules/transactionpool/dump.go
@@ -1,0 +1,3 @@
+package transactionpool
+
+// When doing a dump, make sure that all of the contracts are still legal.
